+++ src/wardrobe_show.py
@@ -15,6 +15,7 @@ from ui.ui_wardrobe_show import Ui_DialogWardrobeShow
 class WardrobeShow(Ui_DialogWardrobeShow, QDialog):
     path_costumes=os.path.expanduser('~')+"/.wardrobe/costumes"
     path_accessories=os.path.expanduser('~')+"/.wardrobe/accessories"
+    path_servers=os.path.expanduser('~')+"/.wardrobe/servers"
 
     def __init__(self, parent=None):
         super().__init__(parent) # parent
@@ -44,14 +45,24 @@ class WardrobeShow(Ui_DialogWardrobeShow, QDialog):
                 found_accessories= os.listdir(self.path_accessories)
                 accessories=sorted(found_accessories)
 
+        servers=[]
+        if os.path.exists(self.path_servers):
+            if os.path.exists(self.path_servers):    
+                found_servers= os.listdir(self.path_servers)
+                servers=sorted(found_servers)
+
         distros=['arch', 'debian', 'ubuntu']
+
+        self.comboBoxDistros.addItems(distros)
+        self.comboBoxDistros.setCurrentIndex(1)
 
         self.comboBoxCostumes.addItems(costumes)
         self.comboBoxAccessories.addItems(accessories)
-        self.comboBoxDistros.addItems(distros)
+        self.comboBoxServers.addItems(servers)
 
         self.pushButtonShowCostume.clicked.connect(self.costume_show)
         self.pushButtonShowAccessory.clicked.connect(self.accessory_show)
+        self.pushButtonShowServer.clicked.connect(self.server_show)
 
 
     ##
@@ -76,6 +87,25 @@ class WardrobeShow(Ui_DialogWardrobeShow, QDialog):
     @QtCore.Slot()
     def accessory_show(self):
         file_path=filename=self.path_accessories + '/' + self.comboBoxAccessories.currentText() + '/'
+        file_name=self.comboBoxDistros.currentText()+ '.yml'
+        filename=file_path+file_name
+        print(filename)
+        if os.path.isfile(filename):
+            dialog=TextEditor()
+            dialog.setWindowTitle(filename)
+            dialog.setFilename(filename)
+            dialog.openFilename()
+            dialog.exec()
+        else:
+            msgBox = QMessageBox(self)
+            msgBox.setText("There is no version for " + self.comboBoxAccessories.currentText())
+            msgBox.exec()
+
+
+    ##
+    @QtCore.Slot()
+    def server_show(self):
+        file_path=filename=self.path_servers + '/' + self.comboBoxServers.currentText() + '/'
         file_name=self.comboBoxDistros.currentText()+ '.yml'
         filename=file_path+file_name
         print(filename)